refactor: rename package mcmin to mcmax
